requirements.txt
Flask==3.0.0
tensorflow==2.20.0
numpy>=1.26.0
Pillow>=10.0.0
scikit-learn>=1.3.0
opencv-python>=4.8.0
matplotlib>=3.7.0
Werkzeug>=3.0.0
